Initial GUI Model Created
The individual GUI pages are ready for wiring the functions for the most part. The only exception is the MainGUI.java file because I need to talk to about the IO stuff in order to figure that out. I still need to connect each of the GUI pages together so it's a cohesive app but that won't take me that long. Most of the current functionality will print to the console if a button is pressed or table element is selected so play around with that if you want. I can walk you guys through the code next time we meet.